:recycle: Refactor tests to use data fixtures
